moving both tests to hardware testing folder
