--- abr-testing/abr_testing/hw_test_scripts/__init__.py
@@ -1,1 +1,0 @@
-"""ABR Hardware Testing Tools."""
